# assignment_4/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Assignment 4 - Auth API",
    "version": "1.0.0",
    "description": "Secure task API with Supabase Auth"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing email or password" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns access token" },
          "400": { "description": "Missing email or password" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Missing, invalid, or expired token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public info, no auth required",
        "responses": {
          "200": { "description": "Public message" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get the logged-in user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile data" },
          "401": { "description": "Missing, invalid, or expired token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Protected dashboard (proves middleware reuse)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard welcome message" },
          "401": { "description": "Missing, invalid, or expired token" }
        }
      }
    }
  }
}